torch-pipeline/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pycocotools>=2.0.7
opencv-python>=4.8.0
numpy>=1.23
Pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
